refactor : ui, bridgemake, domain 패키지 분리
