+++ src/main/java/bridge/Application.java
@@ -1,5 +1,10 @@
 package bridge;
 
+import bridge.domain.Game;
+import bridge.makebridge.BridgeMaker;
+import bridge.makebridge.BridgeNumberGenerator;
+import bridge.ui.InputView;
+
 import java.util.List;
 
 public class Application {
