+++ src/main/java/bridge/domain/BridgeGame.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.domain;
 
 import java.util.ArrayList;
 import java.util.List;
