+++ src/main/java/bridge/makebridge/BridgeMaker.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.makebridge;
 
 import java.util.ArrayList;
 import java.util.List;
