+++ src/main/java/bridge/makebridge/BridgeNumberGenerator.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.makebridge;
 
 @FunctionalInterface
 public interface BridgeNumberGenerator {
