+++ src/main/java/bridge/ui/Validate.java
@@ -1,4 +1,6 @@
-package bridge;
+package bridge.ui;
+
+import bridge.ui.OutputView;
 
 public class Validate {
     OutputView outputView = new OutputView();
